requirements.txt
pyobjc-core ; sys_platform == 'darwin'
pyyaml --global-option="--with-libyaml"
